renamed plugin-build-dev.sh -> plugin-build.sh

+++ plugin-build.sh
@@ -7,16 +7,26 @@ git update-index --assume-unchanged ./plugin-menus.json
 
 #Process each plugin available in the directory
 for plugin_directory in plugins/*/ ; do
+    echo "Visiting $plugin_directory"
+
+    #Skip non-directories 
+    if [ ! -d $plugin_directory ]
+    then
+      echo "Skipping $plugin_directory"
+      continue
+    fi
+
+
     PLUGIN_NAME=$(basename $plugin_directory)
     echo "Processing ${PLUGIN_NAME}..."
     PLUGIN="${plugin_directory}plugin.json"
     PLUGIN_ARRAY="$PLUGIN_ARRAY $PLUGIN"
 
-    #link over the page files to the plugin directories
-    #ln -sf "../../${plugin_directory}/pages" pages/plugins/${PLUGIN_NAME} 
-
     rm -Rf pages/plugins/${PLUGIN_NAME}
-    cp -r ${plugin_directory}/pages pages/plugins/${PLUGIN_NAME}
+    rm -Rf pages/api/plugins/${PLUGIN_NAME}
+    
+    cp -r ${plugin_directory}/pages pages/plugins/${PLUGIN_NAME} || true
+    cp -r ${plugin_directory}/pages/api pages/api/plugins/${PLUGIN_NAME} || true
 done
 
 case $1 in
